Lookup of implementations not limited to sub packages anymore.
Previously implementations had to be stored in the same or in a sub
package of the interface package. 

This is changed now. Now by default the whole classpath is searched for
implementations. This insults performance but can be switched to the old
behavior with new InjectionObjectFactory(new
LookupContext(PackageScope.SUBPACKAGES_ONLY)). Additionally a custom
package can be set with new InjectionObjectFactory(new
LookupContext("com.github.kaiwinter"));